what still needs to be done:
- verify model & converter..registered these issues..won't fix these issues in this PR (since they require changes in both inbound and the engine): https://unileverfoodsolutions.atlassian.net/browse/OHUB-1784,
https://unileverfoodsolutions.atlassian.net/browse/OHUB-1789,
https://unileverfoodsolutions.atlassian.net/browse/OHUB-1790
- add test data for subscription

+++ spark-jobs/src/main/scala/com/unilever/ohub/spark/acm/model/AcmOrderLine.scala
@@ -5,7 +5,7 @@ case class AcmOrderLine(
     ORD_INTEGRATION_ID: String,
     QUANTITY: Long,
     AMOUNT: BigDecimal,
-    LOYALTY_POINTS: Option[Long],
+    LOYALTY_POINTS: Option[BigDecimal],
     PRD_INTEGRATION_ID: String,
     SAMPLE_ID: String,
     CAMPAIGN_LABEL: Option[String],
